requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
Django==4.2
django-filter==24.2
django-seed==0.3.1
djangorestframework==3.15.1
djangorestframework-api-key==3.0.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
Faker==24.9.0
httpie==3.2.2
idna==3.7
iniconfig==2.0.0
Markdown==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
mysqlclient==2.2.4
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pluggy==1.5.0
Pygments==2.18.0
PyJWT==1.7.1
PySocks==1.7.1
pytest==8.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
toposort==1.10
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
